fastapi_service/requirements.txt
fastapi
uvicorn
python-dotenv
pinecone
python-multipart
numpy